Fixes #502, remove parameter dtype (#505)
* Fixes #502, remove dtype parameter
* add warnings for a removed parameter
* Update _topology.py
* Update nearest_neighbours.py
* Update test_sklearn_nearest_neighbour_converter.py

+++ skl2onnx/operator_converters/imputer_op.py
@@ -20,7 +20,7 @@ def convert_sklearn_imputer(scope, operator, container):
     if not hasattr(op, 'statistics_'):
         raise RuntimeError("Member statistics_ is not present, was the "
                            "model fitted?")
-    attrs['imputed_value_floats'] = op.statistics_
+    attrs['imputed_value_floats'] = op.statistics_.astype(np.float32)
     if isinstance(op.missing_values, str) and op.missing_values == 'NaN':
         attrs['replaced_value_float'] = np.NaN
     elif isinstance(op.missing_values, float):
